Updated nginx config creation to match bind config creation
Updated nginx configuration file generation to abide by the domain lists with and without *. prefixes, instead of just appending all domains with . (which matches domain and subdomains).  This was done to keep in-line with the new way bind configuration files are generated, which also now abide by the domain lists with and without *. prefixes.