Updated tailing of log files
Updated tail command to continue tracking files, even if they are removed/renamed, which may occur if the log files are rotated/renamed.  Also changed the tail command to only start logging from the point at which they are executed, thus not including the last 10 lines, which is the tail command default.  This change should result in cleaner startup logs of the image.